7: move vec parse methods to vec_io header
