+++ include/vecmath/vec_ext.h
@@ -22,7 +22,6 @@
 #include "vec.h"
 
 #include <array>
-#include <string_view>
 #include <vector>
 
 namespace vm {
@@ -156,100 +155,6 @@ namespace vm {
     template <typename T, std::size_t S, std::size_t N>
     constexpr std::array<vec<T,S>, N> operator*(const T lhs, const std::array<vec<T,S>, N>& rhs) {
         return rhs * lhs;
-    }
-
-    namespace detail {
-        template <typename T, std::size_t S>
-        bool doParse(const std::string_view& str, size_t& pos, vec<T,S>& result) {
-            constexpr auto blank = " \t\n\r()";
-
-            for (std::size_t i = 0; i < S; ++i) {
-                if ((pos = str.find_first_not_of(blank, pos)) == std::string::npos) {
-                    return false;
-                }
-                result[i] = static_cast<T>(std::atof(str.data() + pos));
-                if ((pos = str.find_first_of(blank, pos)) == std::string::npos) {
-                    if (i < S-1) {
-                        return false;
-                    }
-                }
-            }
-            return true;
-        }
-    }
-
-    /**
-     * Parses the given string representation. The syntax of the given string is as follows
-     *
-     *   VEC ::= S * COMP;
-     *     S ::= number of components
-     *  COMP ::= WS, FLOAT;
-     *    WS ::= " " | \\t | \\n | \\r | "(" | ")";
-     * FLOAT ::= any floating point number parseable by std::atof
-     *
-     * Note that this method does not signal if the string could actually be parsed.
-     *
-     * @tparam T the component type
-     * @tparam S the number of components
-     * @param str the string to parse
-     * @param defaultValue the default value to return if parsing fails
-     * @return the vector parsed from the string
-     */
-    template <typename T, std::size_t S>
-    vec<T,S> parse(const std::string_view& str, const vec<T,S>& defaultValue = vec<T,S>::zero()) {
-        std::size_t pos = 0;
-        vec<T,S> result;
-        if (detail::doParse(str, pos, result)) {
-            return result;
-        } else {
-            return defaultValue;
-        }
-    }
-
-    /**
-     * Returns whether parse() can parse S components from the given string
-     *
-     * @tparam T the component type
-     * @tparam S the number of components
-     * @param str the string to parse
-     * @return whether S components can be parsed
-     */
-    template <typename T, std::size_t S>
-    bool can_parse(const std::string_view& str) {
-        std::size_t pos = 0;
-        vec<T,S> result;
-        return detail::doParse(str, pos, result);
-    }
-
-    /**
-     * Parses the given string for a list of vectors. The syntax of the given string is as follows:
-     *
-     * LIST ::= VEC, { SEP, VEC }
-     *  SEP ::= " " | \\t | \\n |t \\r | "," | ";";
-     *
-     * Note that the list can be separated by whitespace or commas or semicolons, or a mix of these separators. Only
-     * vectors which conform to the vector syntax are added to the given output iterator.
-     *
-     * @tparam T the component type
-     * @tparam S the number of components
-     * @tparam O the type of the output iterator
-     * @param str the string to parse
-     * @param out the output iterator add the parsed vectors to
-     */
-    template <typename T, std::size_t S, typename O>
-    void parse_all(const std::string_view& str, O out) {
-        constexpr auto blank = " \t\n\r,;";
-
-        std::size_t pos = 0;
-        while (pos != std::string::npos) {
-            vec<T,S> temp;
-            if (detail::doParse(str, pos, temp)) {
-                out = temp;
-                ++out;
-            }
-            pos = str.find_first_of(blank, pos);
-            pos = str.find_first_not_of(blank, pos);
-        }
     }
 }
