+++ test/src/vec_ext_test.cpp
@@ -56,81 +56,4 @@ namespace vm {
         CER_ASSERT_EQ(exp, in * 3.0f)
         CER_ASSERT_EQ(exp, 3.0f * in)
     }
-
-    TEST(vec_ext_test, parse_valid_string) {
-        constexpr auto s = "1.0 3 3.5";
-
-        const auto success = can_parse<float, 3>(s);
-        ASSERT_TRUE(success);
-
-        const auto result = parse<float, 3>(s);
-        ASSERT_EQ(vec3f(1.0f, 3.0f, 3.5f), result);
-    }
-
-    TEST(vec_ext_test, parse_short_string) {
-        constexpr auto s = "1.0 3";
-
-        const auto success = can_parse<float, 3>(s);
-        ASSERT_FALSE(success);
-
-        const auto result = parse<float, 3>(s);
-        ASSERT_EQ(vec3f::zero(), result);
-    }
-
-    TEST(vec_ext_test, parse_long_string) {
-        constexpr auto s = "1.0 3 4 5";
-
-        const auto success = can_parse<float, 3>(s);
-        ASSERT_TRUE(success);
-
-        const auto result = parse<float, 3>(s);
-        ASSERT_EQ(vec3f(1.0f, 3.0f, 4.0f), result);
-    }
-
-    TEST(vec_ext_test, parse_invalid_string) {
-        constexpr auto s = "asdf";
-
-        const auto success = can_parse<float, 3>(s);
-        ASSERT_FALSE(success);
-
-        const auto result = parse<float, 3>(s);
-        ASSERT_EQ(vec3f::zero(), result);
-    }
-
-    TEST(vec_ext_test, parse_empty_string) {
-        constexpr auto s = "";
-
-        const auto success = can_parse<float, 3>(s);
-        ASSERT_FALSE(success);
-
-        const auto result = parse<float, 3>(s);
-        ASSERT_EQ(vec3f::zero(), result);
-    }
-
-    TEST(vec_ext_test, parse_all) {
-        std::vector<vec3f> result;
-
-        parse_all<float, 3>("", std::back_inserter(result));
-        ASSERT_TRUE(result.empty());
-
-        result.clear();
-        parse_all<float, 3>("1.0 3 3.5 2.0 2.0 2.0", std::back_inserter(result));
-        ASSERT_EQ(std::vector<vec3f>({ vec3f(1, 3, 3.5), vec3f(2, 2, 2) }), result);
-
-        result.clear();
-        parse_all<float, 3>("(1.0 3 3.5) (2.0 2.0 2.0)", std::back_inserter(result));
-        ASSERT_EQ(std::vector<vec3f>({ vec3f(1, 3, 3.5), vec3f(2, 2, 2) }), result);
-
-        result.clear();
-        parse_all<float, 3>("(1.0 3 3.5), (2.0 2.0 2.0)", std::back_inserter(result));
-        ASSERT_EQ(std::vector<vec3f>({ vec3f(1, 3, 3.5), vec3f(2, 2, 2) }), result);
-
-        result.clear();
-        parse_all<float, 3>("(1.0 3 3.5); (2.0 2.0 2.0)", std::back_inserter(result));
-        ASSERT_EQ(std::vector<vec3f>({ vec3f(1, 3, 3.5), vec3f(2, 2, 2) }), result);
-
-        result.clear();
-        parse_all<float, 3>("1.0 3 3.5, 2.0 2.0 2.0", std::back_inserter(result));
-        ASSERT_EQ(std::vector<vec3f>({ vec3f(1, 3, 3.5), vec3f(2, 2, 2) }), result);
-    }
 }
